Move store and hooks tests
