+++ lib/store/__tests__/useModalStore.test.tsx
@@ -1,5 +1,5 @@
 "use client";
-import useModalStore, { defaultProperties } from "../../../../lib/store/useModalStore";
+import useModalStore, { defaultProperties } from "../useModalStore";
 import { renderHook } from "@testing-library/react";
 import { act } from "react-dom/test-utils";
 import { HTMLTextInputTypeAttribute } from "@lib/types";
