+++ lib/store/__tests__/useTemplateStore.test.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React from "react";
-import { useTemplateStore, TemplateStoreProvider } from "../../../../lib/store/useTemplateStore";
+import { useTemplateStore, TemplateStoreProvider } from "../useTemplateStore";
 import { renderHook, act } from "@testing-library/react";
 
 const createStore = () => {
